+++ RELEASE_NOTES.md
@@ -6,6 +6,10 @@
 ### Bug fixes
 - ...
 
+### Changes
+- Added CLARIN and DARIAH logo's on the main page #87
+- Course Statistics: Added chart of new courses per month #89
+
 ### Small changes
 - Updated dependencies
 - Upgraded CakePHP to 4.5.2